+++ tests/test_package.py
@@ -4,4 +4,4 @@ import dot
 
 
 def test_package_version() -> None:
-    assert dot.__version__ == version("dot") == "1.0.0"
+    assert dot.__version__ == version("dot") == "1.1.0"